Day2/challenge/app/requirements.txt
numpy==1.24.4
pandas==2.1.4
psycopg2-binary==2.9.9
